Bump Rancher dependency to fix EKS launch template issue
There is an issue where EKS launch templates did not get sent to Rancher as
expected. This was because the ID field was missing in the Rancher client. This
has been updated in Rancher and this change brings those changes here.

In addition, there were other changes in the Rancher client. This change brings
in those changes as well (mostly related to ACE config on RKE2 clusters).