fix: restructure machine pool fields
Machine pools are represented in the provisioning cluster resource as a list,
so when an element in the list is removed, the next element takes its place in
the list. For most fields, this works out ok and the fields from the next
element are correctly set in the new list position.

However, for computed fields, if the next element had them set to null (or
omitted), they would take the values from the previous state element that was
removed.

To fix this, this PR removes the computed attribute from these fields: labels,
annotations and machine_labels (which correspond in the cluster CRD to
machineDeploymentLabels, machineDeploymentAnnotations, and labels,
respectively).

These fields aren't actual metadata labels or fields, they are in the cluster
spec and are copied by rancher controllers to labels and metadata of other
objects, so it's normal for them to be just optional, instead of computed +
optional.

The "diffSuppressFunctions" are also removed, since they are mainly useful for
actual metadata labels in other objects.